Update the mission
